projects/dataflow-framework/abstraction-level-8/pyproject.toml
[project]
name = "abstraction-level-8"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "matplotlib>=3.10.6",
    "networkx>=3.5",
    "psutil>=7.0.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "starlette>=0.47.3",
    "typer>=0.17.3",
    "uvicorn>=0.35.0",
]
